refactor: break up fribbles.rs
