+++ src/export/fribbels/mod.rs
@@ -1,0 +1,16 @@
+//! Output format based on the format used by [Fribbels HSR Optimizer],
+//! devised by [kel-z's HSR-Scanner].
+//!
+//! [Fribbels HSR Optimizer]: https://github.com/fribbels/hsr-optimizer
+//! [kel-z's HSR-Scanner]: https://github.com/kel-z/HSR-Scanner
+
+mod converters;
+mod models;
+mod utils;
+mod handlers;
+mod exporter;
+
+pub use models::*;
+pub use converters::*;
+pub use utils::*;
+pub use exporter::*;
